Rename tracecode.tracecode to tracecode.matchers #1
This is better than a name repetition.
Move the TraceCode object to matcher and rename as DeploymentAnalysis
Also perform minor formatting
